+++ src/dxrpy/on_demand_classifier/__init__.py
@@ -1,2 +1,2 @@
 from .job import OnDemandClassifierJob
-from .on_demand_classifier import OnDemandClassifier
+from .on_demand_classifier import OnDemandClassifier, RunJobResult